+++ library/ui/src/main/java/com/google/android/exoplayer2/ui/SubtitleView.java
@@ -271,15 +271,15 @@ public final class SubtitleView extends View implements TextOutput {
 
     for (int i = 0; i < cueCount; i++) {
       Cue cue = cues.get(i);
-      float textSizePx =
-          resolveTextSizeForCue(cue, rawViewHeight, viewHeightMinusPadding, defaultViewTextSizePx);
+      float cueTextSizePx = resolveCueTextSize(cue, rawViewHeight, viewHeightMinusPadding);
       SubtitlePainter painter = painters.get(i);
       painter.draw(
           cue,
           applyEmbeddedStyles,
           applyEmbeddedFontSizes,
           style,
-          textSizePx,
+          defaultViewTextSizePx,
+          cueTextSizePx,
           bottomPaddingFraction,
           canvas,
           left,
@@ -289,14 +289,13 @@ public final class SubtitleView extends View implements TextOutput {
     }
   }
 
-  private float resolveTextSizeForCue(
-      Cue cue, int rawViewHeight, int viewHeightMinusPadding, float defaultViewTextSizePx) {
+  private float resolveCueTextSize(Cue cue, int rawViewHeight, int viewHeightMinusPadding) {
     if (cue.textSizeType == Cue.TYPE_UNSET || cue.textSize == Cue.DIMEN_UNSET) {
-      return defaultViewTextSizePx;
+      return 0;
     }
     float defaultCueTextSizePx =
         resolveTextSize(cue.textSizeType, cue.textSize, rawViewHeight, viewHeightMinusPadding);
-    return defaultCueTextSizePx > 0 ? defaultCueTextSizePx : defaultViewTextSizePx;
+    return Math.max(defaultCueTextSizePx, 0);
   }
 
   private float resolveTextSize(